+++ NEWS.md
@@ -1,4 +1,10 @@
 ### Features and Fixes
+#### version 1.3.1 (11/19/14)
+    * fixed CTS translations
+    * added ID merging for additive translations
+    * added column removal to dataset managment
+    * added package installation of WGCNA and Chemminer, previously missing dependancies
+
 #### version 1.3 (09/23/14)
     * added interactive network visualization using D3.js
     * added network output as svg